requirement.txt
streamlit
transformers
torch
sentencepiece
newspaper3k
requests
beautifulsoup4